requirements.txt
flask
langchain
openai
chromadb
opencc-python-reimplemented
tiktoken
langchain-openai
langchain-community
gunicorn